do neovim stuff

+++ extras/neovim/neovim.sh
@@ -2,9 +2,14 @@
 
 # install neovim
 # sudo apt install neovim
-wget https://github.com/neovim/neovim/releases/download/nightly/nvim-linux64.deb
-sudo apt install ./nvim-linux64.deb
-rm ./nvim-linux64.deb
+echo "Install Neovim?"
+read a
+
+if [[ $a =~ "^y" ]]; then
+	wget https://github.com/neovim/neovim/releases/download/nightly/nvim-linux64.deb
+	sudo apt install ./nvim-linux64.deb
+	rm ./nvim-linux64.deb
+fi
 
 # install vim plug
 sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs \
